logstor: use fixed-width types for disk sizes
for disk size values in the segment_manager use fixed-width types
instead of size_t. size_t should be used only for memory sizes.

+++ replica/logstor/segment_manager.hh
@@ -36,15 +36,15 @@ class compaction_manager;
 class segment_set;
 class primary_index;
 
-static constexpr size_t default_segment_size = 128 * 1024;
-static constexpr size_t default_file_size = 32 * 1024 * 1024;
+static constexpr uint64_t default_segment_size = 128 * 1024;
+static constexpr uint64_t default_file_size = 32 * 1024 * 1024;
 
 /// Configuration for the segment manager
 struct segment_manager_config {
     std::filesystem::path base_dir;
-    size_t segment_size = default_segment_size;
-    size_t file_size = default_file_size;
-    size_t disk_size;
+    uint64_t segment_size = default_segment_size;
+    uint64_t file_size = default_file_size;
+    uint64_t disk_size;
     bool format_on_startup = true;
     bool compaction_enabled = true;
     size_t max_segments_per_compaction = 8;
@@ -130,7 +130,7 @@ public:
     void set_trigger_compaction_hook(std::function<void()> fn);
     void set_trigger_separator_flush_hook(std::function<void(segment_sequence)> fn);
 
-    size_t get_segment_size() const noexcept;
+    uint64_t get_segment_size() const noexcept;
 
     future<> discard_segments(segment_set&);
